sn-provider: add a wrapper around starknet provider
