+++ src/providers/sn_provider/mod.rs
@@ -1,0 +1,3 @@
+pub mod starknet_provider;
+
+pub use starknet_provider::StarknetProvider;
